- Changed: DetailDisplayHandler is now displaying coherent with /item informations (ie Type for /item type field) - Changed: DetailDisplayHandler is including consignment merchant item routine - Bugfix: to avoid identical hashes that causes error logs, type in QuestMgr is now uint instead of ushort previously
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@1136 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/quests/Quests/QuestMgr.cs
@@ -250,7 +250,7 @@ namespace DOL.GS.Quests
         /// <param name="type"></param>
         public static void RegisterQuestType(Type type)
         {
-            ushort typeId =(ushort) type.GetHashCode();
+            uint typeId =(uint) type.GetHashCode();
             if (m_questTypeMap.Contains(typeId))
             {
                 if (log.IsErrorEnabled)
